old shared ptr impl removed && new impl placed
