+++ src/shared_ptr.hpp
@@ -9,355 +9,6 @@
 
 namespace memory {
 
-template<
-    typename T,
-    typename Allocator = std::allocator<T>
->
-class shared_ptr {
-public:
-
-    shared_ptr()
-    : impl(nullptr)
-    { }
-
-    // Take ownership of obj
-    // Argument is pointer to derived object
-    shared_ptr(T* obj)
-    : impl(nullptr) {
-        if(obj != nullptr) {
-            using cb_type = detail::shared_ptr_impl<T, Allocator>;
-            using cb_alloc_traits = std::allocator_traits< typename cb_type::allocator_type >;
-
-            typename cb_type::allocator_type cb_alloc;
-
-            impl = cb_alloc_traits::allocate(cb_alloc, 1);
-            scoped_ptr cb_guard(impl); /*if constructor fails*/
-
-            cb_alloc_traits::construct(cb_alloc, impl, obj);
-
-            scoped_relax(cb_guard);
-        }
-    }
-
-    shared_ptr(const shared_ptr& other)
-    : impl(other.impl) {
-        if(impl != nullptr) {
-            ++impl->ref_count;
-        }
-    }
-
-    ~shared_ptr() {
-        release_ptr();    
-    }
-
-    shared_ptr&
-    operator= (shared_ptr other) {
-        swap(other); // copy and swap
-        
-        return *this;
-    }
-
-    T&
-    operator*() {
-        assert(impl != nullptr);
-        return *impl->obj;
-    }
-
-    const T&
-    operator*() const {
-        assert(impl != nullptr);
-        return *impl->obj;
-    }
-
-    T*
-    operator->() {
-        assert(impl != nullptr);
-        return impl->obj;
-    }
-
-    const T*
-    operator->() const {
-        assert(impl != nullptr);
-        return impl->obj;
-    }
-
-    bool operator==(const shared_ptr& b) const {
-        if(impl == nullptr || b.impl == nullptr)
-            return impl == nullptr && b.impl == nullptr;
-
-        return impl->obj == b.impl->obj;
-    }
-
-    bool operator!=(const shared_ptr& b) const {
-        return !(*this == b);
-    }
-
-    void swap(shared_ptr& other) {
-        using std::swap;
-
-        swap(impl, other.impl);
-    }
-
-    friend void swap(shared_ptr& a, shared_ptr& b) {
-        a.swap(b);
-    }
-
-private:
-
-    void release_ptr() {
-        if(impl != nullptr) {
-            int stored_val = impl->ref_count.load();
-            while(!impl->ref_count.compare_exchange_weak(stored_val, stored_val - 1))
-                ;
-
-            // TODO: CHANGE TO SUPPORT ALLOCATOR
-            if(stored_val - 1 == 0) {
-                using cb_type = detail::shared_ptr_impl<T, Allocator>;
-                using cb_alloc_traits = std::allocator_traits< typename cb_type::allocator_type >;
-
-                typename cb_type::allocator_type cb_alloc;
-
-                cb_alloc_traits::destroy(cb_alloc, impl);
-                cb_alloc_traits::deallocate(cb_alloc, impl, 1);
-            }
-
-            impl = nullptr;
-        }
-    }
-
-    detail::shared_ptr_impl<T, Allocator>* impl;
-};
-
-
-template<typename T, typename Allocator = std::allocator<T>, typename ...Args>
-shared_ptr<T, Allocator> make_shared(Args... args) {
-    using obj_alloc_traits = std::allocator_traits< Allocator >;
-    
-    // TODO: Reduce to inplace cb & obj allocator call
-    Allocator obj_alloc;
-    T *ptr = obj_alloc_traits::allocate(obj_alloc, 1);
-    scoped_ptr obj_guard(ptr, obj_alloc); /*if constructor fails*/
-
-    obj_alloc_traits::construct(obj_alloc, ptr, args...);
-    
-    scoped_relax(obj_guard);
-    return shared_ptr<T, Allocator>(ptr);
-}
-
-
-namespace new_impl {
-
-class sp_cb_base {
-public:
-
-    sp_cb_base()
-    : ref_count(1)
-    {}
-
-    virtual ~sp_cb_base() {}
-
-    void incr_ref()
-    { ++ref_count; }
-
-    void decr_ref() {
-        int stored_val = ref_count.load();
-        while(!ref_count.compare_exchange_weak(stored_val, stored_val - 1))
-            ;
-
-        // TODO: CHANGE TO SUPPORT ALLOCATOR
-        if(stored_val - 1 == 0) {
-            destroy();
-        }
-    }
-
-    virtual void* get_ptr() = 0;
-
-protected:
-    virtual void destroy() = 0;
-
-    sp_cb_base(const sp_cb_base& other) = delete;
-    sp_cb_base& operator=(const sp_cb_base& other) = delete;
-
-private:
-    std::atomic<int> ref_count;
-}; // class sp_cb_base
-
-template<typename T, typename Allocator>
-class sp_cb_separate: public sp_cb_base {
-public:
-
-    typedef 
-        std::allocator_traits<Allocator>::template rebind_alloc<sp_cb_separate>
-        allocator_type;
-
-    sp_cb_separate(T* ptr, const Allocator& alloc_ref)
-    : sp_cb_base()
-    , m_obj(ptr)
-    , m_alloc_ref(alloc_ref)
-    { }
-
-    virtual ~sp_cb_separate() {
-        using obj_alloc_traits = std::allocator_traits<Allocator>;
-        Allocator obj_alloc(m_alloc_ref);
-
-        obj_alloc_traits::destroy(obj_alloc, m_obj);
-        obj_alloc_traits::deallocate(obj_alloc, m_obj, 1);
-    }
-
-    virtual void* get_ptr()
-    { return reinterpret_cast<void*>(m_obj); };
-
-    virtual void destroy() { 
-        using cb_alloc_traits = std::allocator_traits<allocator_type>;
-        allocator_type cb_alloc;
-
-        cb_alloc_traits::destroy(cb_alloc, this);
-        cb_alloc_traits::deallocate(cb_alloc, this, 1);
-    };
-
-
-    sp_cb_separate(const sp_cb_separate& other) = delete;
-    sp_cb_separate& operator=(const sp_cb_separate& other) = delete;
-
-private:
-    T* m_obj;
-    const Allocator& m_alloc_ref;
-}; // class sp_cb_separate
-
-template<typename T, typename Allocator>
-class sp_cb_inplace: public sp_cb_base {
-public:
-
-    typedef 
-        std::allocator_traits<Allocator>::template rebind_alloc<sp_cb_inplace>
-        allocator_type;
-
-    template<typename ...Args>
-    sp_cb_inplace(const Allocator& alloc_ref, Args... args)
-    : sp_cb_base()
-    , m_alloc_ref(alloc_ref) { 
-        using obj_alloc_traits = std::allocator_traits<Allocator>;
-        Allocator obj_alloc(m_alloc_ref);
-        
-        obj_alloc_traits::construct(obj_alloc, get_obj_ptr(), args...);
-    }
-
-    virtual ~sp_cb_inplace() {
-        using obj_alloc_traits = std::allocator_traits<Allocator>;
-        Allocator obj_alloc(m_alloc_ref);
-
-        obj_alloc_traits::destroy(obj_alloc, get_obj_ptr());
-    }
-
-    virtual void* get_ptr()
-    { return reinterpret_cast<void*>(get_obj_ptr()); };
-
-    virtual void destroy() { 
-        using cb_alloc_traits = std::allocator_traits<allocator_type>;
-        allocator_type cb_alloc; 
-
-        cb_alloc_traits::destroy(cb_alloc, this);
-        cb_alloc_traits::deallocate(cb_alloc, this, 1);
-    };
-
-    sp_cb_inplace(const sp_cb_inplace& other) = delete;
-    sp_cb_inplace& operator=(const sp_cb_inplace& other) = delete;
-
-private:
-
-    T* get_obj_ptr()
-    { return reinterpret_cast<T*>(&m_obj_mem); }
-
-    std::aligned_storage_t<sizeof(T), alignof(T)> m_obj_mem;
-    const Allocator& m_alloc_ref;
-
-}; // class sp_cb_inplace
-
-struct sp_cb_inplace_tag_t {};
-constexpr sp_cb_inplace_tag_t sp_cb_inplace_tag;
-
-class sp_refcount {
-public:
-    sp_refcount()
-    : m_cb_ptr(nullptr)
-    {}
-
-    sp_refcount(const sp_refcount& other)
-    : m_cb_ptr(other.m_cb_ptr) {
-        if(m_cb_ptr != nullptr)
-            m_cb_ptr->incr_ref();
-    }
-
-    template<typename T>
-    explicit
-    sp_refcount(T* sep_ptr)
-    : sp_refcount(std::allocator<T>(), sep_ptr)
-    {}
-
-    template<typename T, typename Allocator>
-    explicit
-    sp_refcount(Allocator alloc, T* sep_ptr)
-    : m_cb_ptr(nullptr) {
-        using cb_t = sp_cb_separate<T, Allocator>;
-        using cb_alloc_t = typename cb_t::allocator_type;
-        using alloc_traits = std::allocator_traits< cb_alloc_t >;
-        cb_alloc_t cb_alloc;
-
-        cb_t* cb_ptr = alloc_traits::allocate(cb_alloc, 1);
-        // TODO: exception safety
-        alloc_traits::construct(cb_alloc, cb_ptr, /*args*/ sep_ptr, alloc);
-
-        m_cb_ptr = cb_ptr;
-    }
-
-    template<typename T, typename Allocator, typename ...Args>
-    sp_refcount(T*& out_ptr, Allocator obj_alloc, Args... args) {
-        using cb_t = sp_cb_inplace<T, Allocator>;
-        using cb_alloc_t = typename cb_t::allocator_type;
-        using alloc_traits = std::allocator_traits< cb_alloc_t >;
-        cb_alloc_t cb_alloc;
-
-        cb_t* cb_ptr = alloc_traits::allocate(cb_alloc, 1);
-        // TODO: exception safety
-        alloc_traits::construct(cb_alloc, cb_ptr, /*args*/ obj_alloc, args...);
-
-        m_cb_ptr = cb_ptr;
-        out_ptr = reinterpret_cast<T*>(m_cb_ptr->get_ptr());
-    }
-
-    ~sp_refcount() {
-        if(m_cb_ptr != nullptr)
-            m_cb_ptr->decr_ref();
-    }
-
-    sp_refcount& operator=(sp_refcount other) { 
-        swap(other);
-        return *this;
-    }
-
-    template<typename T>
-    T* get()
-    { 
-        assert(m_cb_ptr != nullptr);
-        return reinterpret_cast<T*>(m_cb_ptr->get_ptr());
-    }
-
-    bool operator==(const sp_refcount& other) const
-    { return m_cb_ptr == other.m_cb_ptr; }
-
-    void swap(sp_refcount& other) {
-        using std::swap;
-
-        swap(m_cb_ptr, other.m_cb_ptr);
-    }
-
-    friend void swap(sp_refcount& a, sp_refcount& b)
-    { a.swap(b); }
-
-    sp_cb_base* m_cb_ptr;
-
-}; // class sp_refcount
-
 template<typename T>
 class shared_ptr {
 public:
@@ -448,7 +99,7 @@ public:
 private:
 
     template<typename Allocator, typename ...Args>
-    shared_ptr(sp_cb_inplace_tag_t, Allocator obj_alloc, Args... args)
+    shared_ptr(detail::sp_cb_inplace_tag_t, Allocator obj_alloc, Args... args)
     : m_refcount(m_ptr, obj_alloc, args...)
     {}
 
@@ -456,21 +107,19 @@ private:
     friend shared_ptr<D> allocate_shared(Allocator obj_alloc, Args... args);
 
     T* m_ptr;
-    sp_refcount m_refcount;
+    detail::sp_refcount m_refcount;
 }; // class shared_ptr
 
 template<typename T, typename Allocator, typename ...Args>
 shared_ptr<T> allocate_shared(Allocator obj_alloc, Args... args) {
-    return shared_ptr<T>(sp_cb_inplace_tag, obj_alloc, args...);
+    return shared_ptr<T>(detail::sp_cb_inplace_tag, obj_alloc, args...);
 }
 
 template<typename T, typename ...Args>
 shared_ptr<T> make_shared(Args... args) {
-    return new_impl::allocate_shared<T>( std::allocator<T>{}, args... );
+    return memory::allocate_shared<T>( std::allocator<T>{}, args... );
 }
 
-
-} // namespace new_impl
 
 } // namespace memory
